Separate influx queries and analytics.
