+++ src/db/collections/analytics/mod.rs
@@ -1,17 +1,16 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(feature = "inx")]
+mod influx;
+
 use std::collections::{HashMap, HashSet};
 
 use decimal::d128;
-#[cfg(feature = "inx")]
-use influxdb::{InfluxDbWriteable, Timestamp};
 use mongodb::{bson::doc, error::Error};
 use serde::{Deserialize, Serialize};
 
 use super::{BlockCollection, OutputCollection};
-#[cfg(feature = "inx")]
-use crate::db::influxdb::{InfluxDb, InfluxDbMeasurement};
 use crate::{
     db::MongoDb,
     types::{
@@ -75,61 +74,6 @@ impl Analytics {
             removed_outputs: Default::default(),
             removed_storage_deposits: Default::default(),
         }
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDb {
-    /// Insert all gathered analytics.
-    pub async fn insert_all_analytics(
-        &self,
-        milestone_timestamp: MilestoneTimestamp,
-        milestone_index: MilestoneIndex,
-        mut analytics: Analytics,
-    ) -> Result<(), influxdb::Error> {
-        self.insert(AddressAnalyticsSchema {
-            milestone_timestamp,
-            milestone_index,
-            analytics: analytics.addresses,
-        })
-        .await?;
-        for (kind, outputs) in analytics.outputs.drain() {
-            self.insert(OutputAnalyticsSchema {
-                milestone_timestamp,
-                milestone_index,
-                kind,
-                analytics: outputs,
-            })
-            .await?;
-        }
-        for (kind, outputs) in analytics.unspent_outputs.drain() {
-            self.insert(OutputAnalyticsSchema {
-                milestone_timestamp,
-                milestone_index,
-                kind,
-                analytics: outputs,
-            })
-            .await?;
-        }
-        self.insert(StorageDepositAnalyticsSchema {
-            milestone_timestamp,
-            milestone_index,
-            analytics: analytics.storage_deposits,
-        })
-        .await?;
-        self.insert(ClaimedTokensAnalyticsSchema {
-            milestone_timestamp,
-            milestone_index,
-            analytics: analytics.claimed_tokens,
-        })
-        .await?;
-        self.insert(MilestoneActivityAnalyticsSchema {
-            milestone_timestamp,
-            milestone_index,
-            analytics: analytics.milestone_activity,
-        })
-        .await?;
-        Ok(())
     }
 }
 
@@ -363,23 +307,6 @@ pub struct AddressAnalyticsSchema {
     pub analytics: AddressAnalytics,
 }
 
-#[cfg(feature = "inx")]
-impl InfluxDbWriteable for AddressAnalyticsSchema {
-    fn into_query<I: Into<String>>(self, name: I) -> influxdb::WriteQuery {
-        Timestamp::from(self.milestone_timestamp)
-            .into_query(name)
-            .add_tag("milestone_index", self.milestone_index)
-            .add_field("total_active_addresses", self.analytics.total_active_addresses)
-            .add_field("receiving_addresses", self.analytics.receiving_addresses)
-            .add_field("sending_addresses", self.analytics.sending_addresses)
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbMeasurement for AddressAnalyticsSchema {
-    const NAME: &'static str = "stardust_addresses";
-}
-
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]
 pub struct OutputAnalytics {
     pub count: u64,
@@ -392,26 +319,6 @@ pub struct OutputAnalyticsSchema {
     pub milestone_index: MilestoneIndex,
     pub kind: String,
     pub analytics: OutputAnalytics,
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbWriteable for OutputAnalyticsSchema {
-    fn into_query<I: Into<String>>(self, name: I) -> influxdb::WriteQuery {
-        Timestamp::from(self.milestone_timestamp)
-            .into_query(name)
-            .add_tag("milestone_index", self.milestone_index)
-            .add_tag("kind", self.kind)
-            .add_field("count", self.analytics.count)
-            .add_field(
-                "total_value",
-                self.analytics.total_value.to_string().parse::<u64>().unwrap(),
-            )
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbMeasurement for OutputAnalyticsSchema {
-    const NAME: &'static str = "stardust_outputs";
 }
 
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]
@@ -428,41 +335,6 @@ pub struct StorageDepositAnalyticsSchema {
     pub milestone_timestamp: MilestoneTimestamp,
     pub milestone_index: MilestoneIndex,
     pub analytics: StorageDepositAnalytics,
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbWriteable for StorageDepositAnalyticsSchema {
-    fn into_query<I: Into<String>>(self, name: I) -> influxdb::WriteQuery {
-        Timestamp::from(self.milestone_timestamp)
-            .into_query(name)
-            .add_tag("milestone_index", self.milestone_index)
-            .add_field("output_count", self.analytics.output_count)
-            .add_field(
-                "storage_deposit_return_count",
-                self.analytics.storage_deposit_return_count,
-            )
-            .add_field(
-                "storage_deposit_return_total_value",
-                self.analytics
-                    .storage_deposit_return_total_value
-                    .to_string()
-                    .parse::<u64>()
-                    .unwrap(),
-            )
-            .add_field(
-                "total_key_bytes",
-                self.analytics.total_key_bytes.to_string().parse::<u64>().unwrap(),
-            )
-            .add_field(
-                "total_data_bytes",
-                self.analytics.total_data_bytes.to_string().parse::<u64>().unwrap(),
-            )
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbMeasurement for StorageDepositAnalyticsSchema {
-    const NAME: &'static str = "stardust_storage_deposits";
 }
 
 impl StorageDepositAnalytics {
@@ -486,21 +358,6 @@ pub struct ClaimedTokensAnalyticsSchema {
     pub milestone_timestamp: MilestoneTimestamp,
     pub milestone_index: MilestoneIndex,
     pub analytics: ClaimedTokensAnalytics,
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbWriteable for ClaimedTokensAnalyticsSchema {
-    fn into_query<I: Into<String>>(self, name: I) -> influxdb::WriteQuery {
-        Timestamp::from(self.milestone_timestamp)
-            .into_query(name)
-            .add_tag("milestone_index", self.milestone_index)
-            .add_field("count", self.analytics.count.to_string().parse::<u64>().unwrap())
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbMeasurement for ClaimedTokensAnalyticsSchema {
-    const NAME: &'static str = "stardust_claimed_tokens";
 }
 
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]
@@ -530,27 +387,4 @@ pub struct MilestoneActivityAnalyticsSchema {
     pub milestone_timestamp: MilestoneTimestamp,
     pub milestone_index: MilestoneIndex,
     pub analytics: MilestoneActivityAnalytics,
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbWriteable for MilestoneActivityAnalyticsSchema {
-    fn into_query<I: Into<String>>(self, name: I) -> influxdb::WriteQuery {
-        Timestamp::from(self.milestone_timestamp)
-            .into_query(name)
-            .add_tag("milestone_index", self.milestone_index)
-            .add_field("count", self.analytics.count)
-            .add_field("transaction_count", self.analytics.transaction_count)
-            .add_field("treasury_transaction_count", self.analytics.treasury_transaction_count)
-            .add_field("milestone_count", self.analytics.milestone_count)
-            .add_field("tagged_data_count", self.analytics.tagged_data_count)
-            .add_field("no_payload_count", self.analytics.no_payload_count)
-            .add_field("confirmed_count", self.analytics.confirmed_count)
-            .add_field("conflicting_count", self.analytics.conflicting_count)
-            .add_field("no_transaction_count", self.analytics.no_transaction_count)
-    }
-}
-
-#[cfg(feature = "inx")]
-impl InfluxDbMeasurement for MilestoneActivityAnalyticsSchema {
-    const NAME: &'static str = "stardust_milestone_activity";
 }
